一旦ビルドOK

+++ src/main/java/org/example/ripository/CustomerRepository.java
@@ -3,6 +3,6 @@ package org.example.ripository;
 import org.example.entity.Customer;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface CustomerRepository extends ReactiveCrudRepository<Customer, Long> {
+public interface CustomerRepository extends ReactiveCrudRepository<Customer, Integer> {
 
 }
